Move SoccerINTLv3 to its own file
